Config/add webpack (#2)
* moved main

* move main.js

* remove browserify; bundle.js now built w/ webpack

* file management

* through "asset mgmt" section

* removed watchify, add webpack-dev-server

* cleaned up console.log

* webpack integrated; all JS and CSS in src/; babel integrated
