+++ package.json
@@ -12,15 +12,23 @@
     "alias": "www.wesley.tech"
   },
   "scripts": {
-    "build": "browserify main.js -o public/bundle.js -v",
+    "build": "webpack",
     "deploy": "now && now alias",
-    "dev": "watchify main.js -o public/bundle.js -v",
-    "start": "http-server"
+    "dev": "webpack-dev-server --open",
+    "start": "http-server",
+    "watch": "webpack --progress --watch"
   },
   "devDependencies": {
-    "browserify": "^14.4.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.1",
+    "css-loader": "^0.28.7",
+    "file-loader": "^1.1.5",
+    "http-server": "^0.10.0",
+    "style-loader": "^0.19.0",
     "watchify": "^3.9.0",
-    "http-server": "^0.10.0"
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.3"
   },
   "dependencies": {}
 }
