+++ public/index.html
@@ -12,7 +12,6 @@
     <link rel="stylesheet" href="https://unpkg.com/tachyons@4.8.1/css/tachyons.min.css"/>
     <link rel="stylesheet"
           href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.min.css">
-    <link rel="stylesheet" href="bundle.css">
 </head>
 <body class="avenir near-white bg-near-black">
 
@@ -299,7 +298,7 @@
 </section>
 
 <!-- let there be JavaScript-->
-<script src="bundle.js"></script>
+<script src="app.bundle.js"></script>
 
 <!-- Google Analytics -->
 <script>
@@ -310,7 +309,7 @@
     ga('create', 'UA-68515019-1', 'auto');
     ga('send', 'pageview');
 </script>
-<script async src='https://www.google-analytics.com/analytics.js'></script>
+<!--<script async src='https://www.google-analytics.com/analytics.js'></script>-->
 <!-- End Google Analytics -->
 </body>
 </html>
