+++ webpack.config.js
@@ -1,0 +1,51 @@
+const path = require('path');
+
+module.exports = {
+    entry: {
+        app: './src/main.js',
+        handlers: './src/js/handlers.js'
+    },
+    devServer: {
+      contentBase: './public'
+    },
+    devtool: 'inline-source-map',
+    output: {
+        filename: '[name].bundle.js',
+        path: path.resolve(__dirname, 'public')
+    },
+    module: {
+        rules: [
+            {
+                test: /\.js$/,
+                exclude: /(node_modules|bower_components)/,
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        presets: ['env']
+                    }
+                }
+            },
+            {
+                test: /\.css$/,
+                use: [
+                    'style-loader',
+                    'css-loader'
+                ]
+            },
+            {
+                test: /\.(png|jpg|jpeg|svg|gif)$/,
+                use: [
+                    'file-loader'
+                ]
+            },
+            {
+                test: /\.html$/,
+                use: [
+                    'html-loader',
+                    'file-loader',
+                    'extract-loader'
+                ]
+            }
+        ]
+    }
+};
